Fixed the linflash troubles...
git-svn-id: svn://svn.code.sf.net/p/amsn/code/trunk@6591 f248a3e1-9e13-0410-b2cf-a4429de6d898

+++ amsn/utils/linux/linflash/flash.h
@@ -12,6 +12,7 @@
 
 // Include files, must include windows.h before tk.h and tcl.h before tk.h
 #include <X11/Xutil.h>
+#include <X11/Xatom.h>
 
 #include <tcl.h>
 #include <tk.h>
@@ -64,7 +65,7 @@ EXTERN int Tk_UnFlashWindow (ClientData clientData,
 			   int objc,
 			   Tcl_Obj *CONST objv[]);
 
-EXTERN int flash_window (Tcl_Interp *interp, Tcl_Obj *CONST objv1, long flash);
+EXTERN int flash_window (Tcl_Interp *interp, Tcl_Obj *CONST objv1, int flash);
